+++ src/main/java/com/jnape/palatable/lambda/adt/Either.java
@@ -1,6 +1,7 @@
 package com.jnape.palatable.lambda.adt;
 
 import com.jnape.palatable.lambda.adt.coproduct.CoProduct2;
+import com.jnape.palatable.lambda.functions.builtin.fn2.Peek;
 import com.jnape.palatable.lambda.functions.specialized.checked.CheckedFn1;
 import com.jnape.palatable.lambda.functions.specialized.checked.CheckedSupplier;
 import com.jnape.palatable.lambda.functor.Applicative;
@@ -171,8 +172,7 @@ public abstract class Either<L, R> implements CoProduct2<L, R, Either<L, R>>, Mo
      * @return the Either, unaltered
      */
     public Either<L, R> peek(Consumer<R> rightConsumer) {
-        return peek(l -> {
-        }, rightConsumer);
+        return Peek.peek(rightConsumer, this);
     }
 
     /**